# perf: Simplify airspace pipeline and improve performance by 21%