--- a/airspace_optimization_performance_summary.md
+++ b/airspace_optimization_performance_summary.md
@@ -1,0 +1,17 @@
+# Airspace Optimization Performance Summary
+
+## Performance Impact Analysis
+*Date: 2025-01-20*
+
+### Executive Summary
+The Int32 coordinate optimization and ClipperData pipeline improvements provide significant performance gains across all dataset sizes, with the most dramatic improvements (29-44%) on small to medium datasets that are most commonly viewed by users.
+
+## Overall Performance Improvements
+
+| **Dataset Size** | **Unoptimized** | **Int32** | **Improvement** |
+|---|---|---|---|
+| Tiny (9 polygons) | 93ms | 64ms | **31% faster** |
+| Small (39 polygons) | 284ms | 158ms | **44% faster** |
+| Medium (1022 polygons) | 994ms | 701ms | **29% faster** |
+| Large (1344 polygons) | 1851ms | 1682ms | **9% faster** |
+