--- a/free_flight_log_app/lib/data/models/airspace_cache_models.dart
+++ b/free_flight_log_app/lib/data/models/airspace_cache_models.dart
@@ -6,8 +6,7 @@ class CachedAirspaceGeometry {
   final String id;
   final String name;
   final int typeCode;  // Store numeric type code instead of string
-  final List<List<LatLng>>? polygons;  // Made optional - either polygons or clipperData
-  final ClipperData? clipperData;  // Alternative to polygons for direct clipping
+  final ClipperData clipperData;  // Always use ClipperData for optimal performance
   final Map<String, dynamic> properties;
   final DateTime fetchTime;
   final String geometryHash;
@@ -19,73 +18,19 @@ class CachedAirspaceGeometry {
     required this.id,
     required this.name,
     required this.typeCode,
-    this.polygons,  // Now optional
-    this.clipperData,  // New optional field
+    required this.clipperData,
     required this.properties,
     required this.fetchTime,
     required this.geometryHash,
     this.compressedSize = 0,
     this.uncompressedSize = 0,
     this.lowerAltitudeFt,
-  }) : assert(polygons != null || clipperData != null, 'Either polygons or clipperData must be provided');
+  });
 
-  factory CachedAirspaceGeometry.fromJson(Map<String, dynamic> json) {
-    return CachedAirspaceGeometry(
-      id: json['id'] as String,
-      name: json['name'] as String,
-      typeCode: json['typeCode'] as int? ?? json['type'] as int? ?? 0,  // Handle old 'type' field for compatibility
-      polygons: _parsePolygons(json['polygons']),
-      properties: Map<String, dynamic>.from(json['properties'] ?? {}),
-      fetchTime: DateTime.parse(json['fetchTime'] as String),
-      geometryHash: json['geometryHash'] as String,
-      compressedSize: json['compressedSize'] ?? 0,
-      uncompressedSize: json['uncompressedSize'] ?? 0,
-      lowerAltitudeFt: json['lowerAltitudeFt'] as int?,
-    );
-  }
+  // fromJson and toJson methods removed - data loaded directly from database
 
-  Map<String, dynamic> toJson() {
-    return {
-      'id': id,
-      'name': name,
-      'typeCode': typeCode,  // Store as numeric type code
-      'polygons': polygons != null ? _encodePolygons(polygons!) : null,
-      'properties': properties,
-      'fetchTime': fetchTime.toIso8601String(),
-      'geometryHash': geometryHash,
-      'compressedSize': compressedSize,
-      'uncompressedSize': uncompressedSize,
-      'lowerAltitudeFt': lowerAltitudeFt,
-    };
-  }
 
-  static List<List<LatLng>> _parsePolygons(dynamic data) {
-    if (data is List) {
-      return data.map((polygon) {
-        if (polygon is List) {
-          return polygon.map((point) {
-            if (point is List && point.length >= 2) {
-              // Safely convert to double, handling both int and double values
-              final lat = (point[1] is int) ? point[1].toDouble() : point[1] as double;
-              final lng = (point[0] is int) ? point[0].toDouble() : point[0] as double;
-              return LatLng(lat, lng);
-            }
-            return LatLng(0, 0);
-          }).toList();
-        }
-        return <LatLng>[];
-      }).toList();
-    }
-    return [];
-  }
-
-  static List<List<List<double>>> _encodePolygons(List<List<LatLng>> polygons) {
-    return polygons.map((polygon) {
-      return polygon.map((point) {
-        return [point.longitude, point.latitude];
-      }).toList();
-    }).toList();
-  }
+  // Helper methods removed - using ClipperData for all geometry operations
 
   bool get isExpired {
     final age = DateTime.now().difference(fetchTime);