--- a/free_flight_log_app/lib/services/airspace_geojson_service.dart
+++ b/free_flight_log_app/lib/services/airspace_geojson_service.dart
@@ -403,7 +403,6 @@ class AirspaceGeoJsonService {
   /// Fetch airspace data as GeoJSON string (deprecated - use fetchAirspacePolygonsDirect)
   @Deprecated('Use fetchAirspacePolygonsDirect for better performance')
   Future<String> fetchAirspaceGeoJson(fm.LatLngBounds bounds) async {
-    // Only support country mode now - tile mode removed
     final hasCountries = await hasLoadedCountries();
     if (!hasCountries) {
       LoggingService.warning('No countries loaded, returning empty GeoJSON');
@@ -563,14 +562,6 @@ class AirspaceGeoJsonService {
     return mergedGeoJson;
   }
 
-  /// Fetch airspace data from OpenAIP Core API with hierarchical caching (tile mode)
-  @Deprecated('Tile-based fetching is no longer supported - use country mode')
-  Future<String> _fetchAirspaceFromTiles(fm.LatLngBounds bounds) async {
-    // Tile-based fetching has been removed - only country mode is supported
-    throw UnsupportedError('Tile-based fetching is no longer supported. Please download country data.');
-  }
-
-  // [REMOVED: All tile-based helper methods below - they are no longer used]
 
   /// Convert OpenAIP response format to standard GeoJSON
   String _convertToGeoJson(String responseBody) {
@@ -1929,8 +1920,6 @@ class AirspaceGeoJsonService {
       'performance': metrics,
       'summary': {
         'total_unique_airspaces': stats.totalGeometries,
-        'total_tiles_cached': stats.totalTiles,
-        'empty_tiles': stats.emptyTiles,
         'database_size_mb': databaseSizeMb.toStringAsFixed(2),
         'database_version': dbVersion,
         'memory_saved_mb': (stats.memoryReductionPercent * stats.totalMemoryBytes / 100 / 1024 / 1024).toStringAsFixed(2),